ci: revert to a simpler model of always requiring a version bump
This can be slightly more confusing for contributors fixing just a typo, but is also a much simpler system that can easily fit in ones head and can be overriden when necessary. The alternative is a complex set of rules checking branches, tags, etc.

There is also (debatable) value in forcing CD for documentation. This may seem overkill for a typo fix, but it is hard to draw a line, and the cost is negligible (this project does not build extensions so PyPi wheels are small).